Modify TilePath to use cog endpoint for non-mosiac rasters
