+++ spec/values/tile_path_spec.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+require "rails_helper"
+
+describe TilePath do
+  describe "tilejson" do
+    context "with a MapSet that has Raster grandchildren" do
+      it "returns a mosaic tilejson path" do
+        scanned_map = FactoryBot.create_for_repository(:scanned_map_with_raster_children)
+        map_set = FactoryBot.create_for_repository(:scanned_map, member_ids: [scanned_map.id], id: "331d70a5-4bd9-4a65-80e4-763c8f6b34fd")
+        expect(described_class.new(map_set).tilejson).to eq "https://map-tiles-test.example.com/mosaicjson/tilejson.json?id=331d70a54bd94a6580e4763c8f6b34fd"
+      end
+    end
+
+    context "with a RasterResouce with a GeoTiff FileSet" do
+      it "returns a cog tilejson path" do
+        file_set = FactoryBot.create_for_repository(:geo_raster_cloud_file)
+        raster = FactoryBot.create_for_repository(:raster_resource, member_ids: [file_set.id], id: "331d70a5-4bd9-4a65-80e4-763c8f6b34fd")
+        expect(described_class.new(raster).tilejson).to eq "https://map-tiles-test.example.com/cog/tilejson.json?id=331d70a54bd94a6580e4763c8f6b34fd"
+      end
+    end
+  end
+end
